@@ -1,4 +1,4 @@
-﻿namespace Opti
+﻿namespace Opti.Parser
 {
     using System;
 
@@ -10,11 +10,16 @@
 
         public int Second { get; }
 
-        public GsaLine(string name, int index, int first, int second) : base(name)
+        public GsaLine(string instruction, int index, int first, int second) : base(instruction)
         {
             this.Index = index;
             this.First = first;
             this.Second = second;
+        }
+
+        public override string ToString()
+        {
+            return $"GsaLine: [  {this.Index}\t{this.Instruction}\t{this.First}\t{this.Second}]";
         }
 
         public static GsaLine Parse(string line)
@@ -23,5 +28,10 @@
 
             return new GsaLine(split[1], int.Parse(split[0]), int.Parse(split[2]), int.Parse(split[3]));
         }
+
+        //public static string Make(string instruction, int index, int first, int second)
+        //{
+        //    return $"{$"   {index} {instruction}",-12}{first,-7}{second}";
+        //}
     }
 }
