@@ -3,70 +3,64 @@
     using System;
     using System.Collections.Generic;
     using System.CommandLine;
-    using System.CommandLine.Invocation;
-    using System.Diagnostics;
     using System.IO;
-    using System.Linq;
-    using System.Management;
-    using System.Runtime.InteropServices;
-    
-    using Opti.Properties;
+    using System.Threading.Tasks;
 
-    using Process = System.Diagnostics.Process;
+    using Resources = Properties.Resources;
 
     /// <summary>
     /// The <c>Program</c> class. Handles all input as well as the IO.
     /// </summary>
     internal static class Program
     {
-        private static readonly Lazy<string> Desktop = new Lazy<string>(() => Environment.GetFolderPath(Environment.SpecialFolder.Desktop));
+        private const string DEFAULT_OUTDIR = "ASM_Optimized";
 
-        private static int Main(string[] args)
+        private const string DEFAULT_RESULTNAME = "Result";
+
+        private const string DEFAULT_SEARCHFOLDER = "Studio";
+
+        private static async Task<int> Main(string[] args)
         {
-#if DEBUG
-            Console.Write("opti: ");
-            Print(ConsoleColor.Cyan, "running in DEBUG mode");
-#endif
+            var paths = new Argument<string[]>("paths", Resources.description_paths);
+            var outdir = new Option<string>("--outdir", GetDefaultOutdir, Resources.description_outdir);
+            var resultName = new Option<string>("--resultName", GetDefaultResultName, Resources.description_resultname);
+            var print = new Option<bool>("--print", Resources.description_print);
 
-            var command = new RootCommand(Resources.description_root)
-                              {
-                                  new Argument<string[]>("paths", Resources.description_paths),
-                                  new Option<bool>("--verifyonly", Resources.description_verifyonly),
-                                  new Option<string>(new[] { "--outdir", "--output" }, GetDefaultOutdir, Resources.description_outdir),
-                                  new Option<string>("--resultName", GetDefaultResultName, Resources.description_resultname),
-                                  new Option<bool>("--print", Resources.description_print),
-                                  new Option<bool>("--open", Resources.description_open)
-                              };
+            var command = new RootCommand(Resources.description_root) { paths, outdir, resultName, print };
+            command.SetHandler((Action<string[], string, string, bool>)InvokeOptimizer, paths, outdir, resultName, print);
 
-            command.Handler = CommandHandler.Create<string[], bool, string, string, bool, bool>(InvokeOptimizer);
+            var result = await command.InvokeAsync(args);
 
-            var result = command.Invoke(args);
-
-            if (result != 0 && IsDisappearing())
+            if (result != 0)
             {
-                Console.WriteLine(Resources.press_to_continue);
+                Info(Resources.press_to_continue);
                 Console.ReadKey(true);
             }
 
             return result;
         }
 
-        private static string GetDefaultOutdir() => Path.Combine(Desktop.Value, "ASM_Optimized");
+        private static readonly Lazy<string> Desktop = new(() => Environment.GetFolderPath(Environment.SpecialFolder.Desktop));
 
-        private static string GetDefaultResultName() => "Result";
+        private static string GetDefaultOutdir() => Path.Combine(Desktop.Value, DEFAULT_OUTDIR);
+
+        private static string GetDefaultResultName() => DEFAULT_RESULTNAME;
 
         private static string GetDefaultSearchFolder()
         {
 #if !DEBUG
             return Directory.GetCurrentDirectory();
 #else
-            return Path.Combine(Desktop.Value, "Studio");
+            return Path.Combine(Desktop.Value, DEFAULT_SEARCHFOLDER);
 #endif
         }
 
-        private static int InvokeOptimizer(string[] paths, bool verifyonly, string outdir, string resultName, bool print, bool open)
+        private static void InvokeOptimizer(string[] paths, string outdir, string resultName, bool print)
         {
-            paths ??= new[] { GetDefaultSearchFolder() };
+            if (paths.Length == 0)
+            {
+                paths = new[] { GetDefaultSearchFolder() };
+            }
 
             var files = new List<string>();
 
@@ -107,88 +101,59 @@
 
             if (gsa == null || txt == null || mic == null)
             {
-                return -1;
+                return;
             }
 
             var optimizer = new AsmOptimizer(gsa, txt, mic, resultName);
-            Console.WriteLine(optimizer.IsInputValid ? Resources.input_ok : Resources.input_err);
 
-            if (!optimizer.IsInputValid)
+            if (!optimizer.FileCollection.VerifyStructure())
             {
-                return -1;
+                Info(Resources.input_err);
+                return;
             }
 
-            if (!verifyonly)
+            Info(Resources.input_ok);
+
+            try
             {
-                try
+                Directory.CreateDirectory(outdir);
+            }
+            catch
+            {
+                Error(Resources.err_outdir_invalid);
+                return;
+            }
+
+            Info(Resources.info_optimized, optimizer.Optimize());
+
+            var name = optimizer.GetResultName(outdir);
+
+            if (print)
+            {
+                foreach (var file in optimizer.FileCollection.Files())
                 {
-                    Directory.CreateDirectory(outdir);
-                }
-                catch
-                {
-                    Error(Resources.err_outdir_invalid);
-                    return -1;
-                }
-
-                Console.WriteLine(Resources.info_optimized, optimizer.Optimize());
-
-                if (print)
-                {
-                    var name = optimizer.GetResultName(outdir);
-
-                    foreach (var file in optimizer.Coordinator.Files())
-                    {
-                        var fileName = name + file.Extension;
-                        Console.WriteLine(Resources.info_file, fileName);
-                        Print(ConsoleColor.DarkYellow, string.Join(Environment.NewLine, file.GetContent()));
-                        Console.WriteLine();
-                    }
-                }
-
-                optimizer.SaveTo(outdir);
-                Console.WriteLine(Resources.info_saved, outdir);
-
-                if (open)
-                {
-                    OpenExplorer(outdir);
+                    var fileName = name + file.Extension;
+                    Info(Resources.info_file, fileName);
+                    Print(ConsoleColor.DarkYellow, string.Join(Environment.NewLine, file.GetContent()));
+                    Console.WriteLine();
                 }
             }
 
-            return 0;
-        }
-
-        private static void Print(ConsoleColor color, string message, params object[] arguments)
-        {
-            var tmp = Console.ForegroundColor;
-            Console.ForegroundColor = color;
-            Console.WriteLine(message, arguments);
-            Console.ForegroundColor = tmp;
+            optimizer.SaveTo(outdir);
+            Info(Resources.info_saved, outdir, name);
         }
 
         public static void Error(string message, params object[] arguments) => Print(ConsoleColor.Red, "E: " + message, arguments);
 
         public static void Info(string message, params object[] arguments) => Print(ConsoleColor.Yellow, "I: " + message, arguments);
 
-        private static void OpenExplorer(string directory)
+        private static void Print(ConsoleColor color, string message, params object[] arguments)
         {
-            Process.Start(new ProcessStartInfo { FileName = directory, Verb = "open", UseShellExecute = true });
-        }
+            var tmp = Console.ForegroundColor;
 
-        private static bool IsDisappearing()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                var query = "SELECT ParentProcessId FROM Win32_Process WHERE ProcessId = " + Process.GetCurrentProcess().Id;
-                var mos = new ManagementObjectSearcher(@"root\CIMV2", query);
-
-                var processId = (int)(uint)mos.Get().Cast<ManagementBaseObject>().First()["ParentProcessId"];
-                var processName = Process.GetProcessById(processId).ProcessName;
-
-                Console.WriteLine(processName);
-                return processName == "explorer";
-            }
-
-            return true;
+            Console.ForegroundColor = color;
+            Console.WriteLine(message, arguments);
+            Console.ForegroundColor = tmp;
         }
     }
 }
