@@ -19,7 +19,7 @@ namespace Opti.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -160,7 +160,7 @@ namespace Opti.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Optimized {0} elements.
+        ///   Looks up a localized string similar to Optimized {0} elements total.
         /// </summary>
         internal static string info_optimized {
             get {
@@ -169,7 +169,7 @@ namespace Opti.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Optimization results saved to {0}.
+        ///   Looks up a localized string similar to Optimization results saved to {0} as {1}.*.
         /// </summary>
         internal static string info_saved {
             get {
